+++ Node2/servo.c
@@ -9,7 +9,7 @@
 #include "servo.h"
 
 void servo_set_angle(float angle){
-	int ms = angle; 
+	float ms = angle; 
 	
 	if(ms > 2.1){
 		ms = 2.1;
@@ -17,7 +17,6 @@ void servo_set_angle(float angle){
 	else if (ms < 0.9){
 		ms = 0.9;
 	}
-	
 	pwm_set_duty_cycle(ms);
 	
 	//TODO: map angle to ms
